requirements.txt
cv2
pytesseract
Pillow
typer
